Moved OptionalHostnamePort to non-public package
OptionalHostnamePort was declared as a public class in the public
org.neo4j.helpers public package. This is really an implementation
detail and doesn't need to be public.
The class has been moved to org.neo4j.kernel.impl.util which isn't
public.
