+++ community/kernel/src/main/java/org/neo4j/kernel/impl/util/OptionalHostnamePort.java
@@ -17,10 +17,12 @@
  * You should have received a copy of the GNU General Public License
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
-package org.neo4j.helpers;
+package org.neo4j.kernel.impl.util;
 
 import java.util.Optional;
 import javax.annotation.Nullable;
+
+import org.neo4j.helpers.HostnamePort;
 
 public class OptionalHostnamePort
 {
