+++ enterprise/backup/src/main/java/org/neo4j/backup/BackupStrategy.java
@@ -21,7 +21,7 @@ package org.neo4j.backup;
 
 import java.io.File;
 
-import org.neo4j.helpers.OptionalHostnamePort;
+import org.neo4j.kernel.impl.util.OptionalHostnamePort;
 import org.neo4j.kernel.configuration.Config;
 import org.neo4j.kernel.lifecycle.Lifecycle;
