+++ enterprise/backup/src/main/java/org/neo4j/backup/BackupStrategyWrapper.java
@@ -22,7 +22,7 @@ package org.neo4j.backup;
 import java.io.File;
 
 import org.neo4j.commandline.admin.CommandFailed;
-import org.neo4j.helpers.OptionalHostnamePort;
+import org.neo4j.kernel.impl.util.OptionalHostnamePort;
 import org.neo4j.io.pagecache.PageCache;
 import org.neo4j.kernel.configuration.Config;
 import org.neo4j.kernel.lifecycle.LifeSupport;
