+++ enterprise/backup/src/main/java/org/neo4j/backup/HaBackupStrategy.java
@@ -23,7 +23,7 @@ import java.io.File;
 
 import org.neo4j.com.ComException;
 import org.neo4j.helpers.HostnamePort;
-import org.neo4j.helpers.OptionalHostnamePort;
+import org.neo4j.kernel.impl.util.OptionalHostnamePort;
 import org.neo4j.kernel.configuration.Config;
 import org.neo4j.kernel.impl.store.MismatchingStoreIdException;
 import org.neo4j.kernel.lifecycle.LifecycleAdapter;
